Fixed unpredictable log macros
An issue exists in that the log level can radically change the behaviour
of specific code. The logging macros insert a {} enclosed block of code
when within log level and do nothing otherwise. This makes code like

    else
        WARNING_LOG("Oh noes!!")

unpredictable since there is the potential for nothing to be inserted
after the else. To solve this the macros now insert empty {} when
outside of log level.