Fixed named sockets bug
Manager::listen( ) offers an override for AF_UNIX sockets, with optional user,
group and permissions.  The bug is that the implementation uses fchown() and
fchmod(), which don't work on AF_UNIX socket file descriptors.  What is needed
is to change the file to which the socket is bound.

The implementation can be made to work by using chown()/chmod() on the name of
the socket.  As you first delete the file before the bind() call, I don't think
the security issues with chown()/chmod() are a real concern.